Adds ability to inline files
The value `{{fileContents}}` in the replace template is now replaced
with the contents of the file with EOL characters converted to match the
enclosing file and indented to align with the placeholder.

The indentation is merely prepended to each line of the file. The
indentation from the inlined file is left in tact. Would be nice to
convert that to match the enclosing file as well though.